wip: cleanup queue + destructive-write investigation (F6/F7 specs RED by design)
- MessageBox titles: ASCII hyphen; clang-21 implicit int->float warnings
  fixed (main.cpp, spectral_view.cpp); init.sh fails on make errors instead
  of reporting CLEAN.
- WriteAndRefresh writes the SOURCE format (ReadWavHeader) instead of the
  standalone session's stale m_wavBitsPerSample (default 16): a 24-bit item
  stays 24-bit on Reverse/DC Remove.
- WriteAndRefresh releases SneakPeak's take accessors (live, retained,
  loader) around the write - harmless, but NOT the fix for F7.
- Specs: test_destructive_long (5-min item must keep 44.1k = F6, reversal
  audible in REAPER = F7) and the gates spec driver (native modal dismissed
  with a CGEvent Return, in-REAPER defer trigger). Both RED on purpose:
  F6 = ops write the downsampled buffer at the reduced rate; F7 = tmp+rename
  gives a new inode and REAPER's per-path decoder pool keeps serving the old
  one. Plan: .harness/plan_v250_milestone.md increments 8b/8c.

+++ src/main.cpp
@@ -612,7 +612,7 @@ REAPER_PLUGIN_DLL_EXPORT int ReaperPluginEntry(
     if (us && us[0]) {
       char* e = nullptr;
       long v = strtol(us, &e, 10);
-      if (e != us && v >= 800 && v <= 2000) g_uiScale = v / 1000.0;
+      if (e != us && v >= 800 && v <= 2000) g_uiScale = (double)v / 1000.0;
     }
   }
